+++ lib/lighthouseProvider/devices/LighthouseV2Device.dart
@@ -3,7 +3,9 @@ import 'dart:async';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_blue/flutter_blue.dart';
 import 'package:lighthouse_pm/lighthouseProvider/LighthousePowerState.dart';
+import 'package:lighthouse_pm/lighthouseProvider/ble/DefaultCharacteristics.dart';
 import 'package:lighthouse_pm/lighthouseProvider/devices/BLEDevice.dart';
+import 'package:lighthouse_pm/lighthouseProvider/helpers/FlutterBlueExtensions.dart';
 
 ///The bluetooth service that handles the power state of the device.
 final Guid _POWER_SERVICE = Guid('00001523-1212-efde-1523-785feabcd124');
@@ -15,6 +17,26 @@ class LighthouseV2Device extends BLEDevice {
   LighthouseV2Device(BluetoothDevice device) : super(device);
 
   BluetoothCharacteristic /* ? */ _characteristic;
+  String /* ? */ _firmwareVersion;
+  final Map<String, String> _otherMetadata = Map();
+  static const _SUPPORTED_CHARACTERISTICS_LIST = const [
+    DefaultCharacteristics.MODEL_NUMBER_STRING_CHARACTERISTIC,
+    DefaultCharacteristics.SERIAL_NUMBER_STRING_CHARACTERISTIC,
+    DefaultCharacteristics.HARDWARE_REVISION_CHARACTERISTIC,
+    DefaultCharacteristics.MANUFACTURER_NAME_CHARACTERISTIC,
+  ];
+
+  @override
+  String get firmwareVersion {
+    if (_firmwareVersion == null) {
+      return "UNKNOWN";
+    } else {
+      return _firmwareVersion;
+    }
+  }
+
+  @override
+  Map<String, String> get otherMetadata => _otherMetadata;
 
   @override
   Future<int /*?*/ > getCurrentState() async {
@@ -68,16 +90,42 @@ class LighthouseV2Device extends BLEDevice {
     final List<BluetoothService> services =
         await this.device.discoverServices();
     for (final service in services) {
-      if (service.uuid != _POWER_SERVICE) {
-        continue;
-      }
       // Find the correct characteristic.
       for (final characteristic in service.characteristics) {
+        final uuid = characteristic.uuid.toLighthouseGuid();
+
         if (characteristic.uuid == _POWER_CHARACTERISTIC) {
           this._characteristic = characteristic;
-          return true;
+        }
+        if (DefaultCharacteristics.FIRMWARE_REVISION_CHARACTERISTIC
+            .isEqualToGuid(uuid)) {
+          this._firmwareVersion = await characteristic.readString();
+          this._firmwareVersion =
+              this._firmwareVersion.replaceAll('\r', '').replaceAll('\n', ' ');
+        }
+        for (final defaultCharacteristic in _SUPPORTED_CHARACTERISTICS_LIST) {
+          if (defaultCharacteristic.isEqualToGuid(uuid)) {
+            String response;
+            switch (defaultCharacteristic.type) {
+              case int:
+                final responseInt = await characteristic.readUint32();
+                response = "$responseInt";
+                break;
+              case String:
+                response = await characteristic.readString();
+                break;
+              default:
+                debugPrint('Unsupported type ${defaultCharacteristic.type}');
+            }
+            if (response != null) {
+              _otherMetadata[defaultCharacteristic.name] = response;
+            }
+          }
         }
       }
+    }
+    if (this._characteristic != null) {
+      return true;
     }
     await disconnect();
     return false;
@@ -95,16 +143,12 @@ class LighthouseV2Device extends BLEDevice {
   @override
   Future internalChangeState(LighthousePowerState newState) async {
     if (this._characteristic != null) {
-      switch(newState) {
+      switch (newState) {
         case LighthousePowerState.ON:
-          await this
-              ._characteristic
-              .write([0x01], withoutResponse: true);
+          await this._characteristic.write([0x01], withoutResponse: true);
           break;
         case LighthousePowerState.STANDBY:
-          await this
-              ._characteristic
-              .write([0x00], withoutResponse: true);
+          await this._characteristic.write([0x00], withoutResponse: true);
       }
     }
   }